pyproject.toml
[project]
name = "copilot-sdk-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "github-copilot-sdk>=0.2.2",
]
